edit multiple files; 함수와 클래스를 분리 후 router 설정
